in order to provide window size config, read Maes config info earlier...
which in turn demanded a retreat to regular org.json (v. p5's version),
since we needed to use it before processing's "path" was ready/available.

also demanded some refactoring of PZVivify.

Later, we should relocate the JSON file-read-and-parse methods (borrowed
from fatcrayon and plopped down in PZApplet) to a 'utils' class.