take the giant pile of gradle mashed potato and fashion a wee mountain...
... as R. Dreyfuss teaches.

No, but seriously: this aggregated commit enables gradle builds of
p5zeugma (which, from maven's perch, is called 'zeugma-processing').
Running
  ./gradlew publishToMavenLocal
or any of the publishing variants creates both an a la carte jar
as well as an uberjar; the latter pulls in zeugma-core, osc and json
stuff, processing's java libraries, and processing's native libraries
(for all platforms). We imagine that it is easiest and, thus, preferred
for developers to 'link against' this monolith jar rather than against
the profusion of its components individually.